+++ src/main/java/org/atomspace/camel/component/tinkerforge/device/DistanceIREndpoint.java
@@ -57,8 +57,8 @@ public class DistanceIREndpoint extends TinkerforgeEndpoint<DistanceIRConsumer, 
     private Long period;
     private Long period2;
     private Character option;
-    private Short min;
-    private Short max;
+    private Integer min;
+    private Integer max;
     private Character option2;
     private Integer min2;
     private Integer max2;
@@ -149,8 +149,8 @@ public class DistanceIREndpoint extends TinkerforgeEndpoint<DistanceIRConsumer, 
             case "setDistanceCallbackThreshold":
                 device.setDistanceCallbackThreshold(
                         getValue(char.class, "option", m, getOption()),
-                        getValue(short.class, "min", m, getMin()),
-                        getValue(short.class, "max", m, getMax())
+                        getValue(int.class, "min", m, getMin()),
+                        getValue(int.class, "max", m, getMax())
                     );
                 break;
 
@@ -351,11 +351,11 @@ public class DistanceIREndpoint extends TinkerforgeEndpoint<DistanceIRConsumer, 
      * The default value is ('x', 0, 0).
      * 
      */
-    public Short getMin(){
+    public Integer getMin(){
         return min;
     }
 
-    public void setMin(Short min){
+    public void setMin(Integer min){
         this.min = min;
     }
 
@@ -378,11 +378,11 @@ public class DistanceIREndpoint extends TinkerforgeEndpoint<DistanceIRConsumer, 
      * The default value is ('x', 0, 0).
      * 
      */
-    public Short getMax(){
+    public Integer getMax(){
         return max;
     }
 
-    public void setMax(Short max){
+    public void setMax(Integer max){
         this.max = max;
     }
